refactor(Profit): 미사용 service, repository 클래스 삭제
